build: bump rust-version to 1.85.0 and edition to 2024

+++ crates/memospot/build.rs
@@ -100,10 +100,10 @@ fn cleanup_dummy_deps() {
     if !memos_bin.exists() {
         return;
     }
-    if let Ok(meta) = memos_bin.metadata() {
-        if meta.len() == 0 {
-            fs::remove_file(memos_bin).ok();
-        }
+    if let Ok(meta) = memos_bin.metadata()
+        && meta.len() == 0
+    {
+        fs::remove_file(memos_bin).ok();
     }
 }
